dep/dep-requirements.txt
numpy==1.23.4
torch==1.12.1
torchvision==0.13.1
torchaudio==0.12.1
mmcv-full==1.7.1
mmsegmentation==0.29.1
opencv-python==4.10.0.84
imageio>=2.22.0
Pillow>=9.0
scipy>=1.13.1
